added latest for data type messages
