+++ OpenStockApp.Core.Maui/Contracts/Services/IAndroidNotificationBuilder.cs
@@ -1,0 +1,11 @@
+﻿#if ANDROID
+using Android.App;
+
+namespace OpenStockApp.Core.Maui.Contracts.Services
+{
+    public interface IAndroidNotificationBuilder
+    {
+        Notification CreateNotification(OpenStockApi.Core.Models.Searches.Result result);
+    }
+}
+#endif
